+++ tests/_core/metrics/_distance/conftest.py
@@ -13,7 +13,7 @@ NAMED_METRICS = (
     DistanceMetric.linf_chebyshev(),
     DistanceMetric.cosine(),
     DistanceMetric.geometric_mean(),
-    DistanceMetric.lminusinf_min_coordinate(),
+    DistanceMetric.l_minus_inf(),
     DistanceMetric.along_axis(1),
 )